requirements.txt
fastapi
uvicorn
google-generativeai
requests
python-dotenv
pydantic
openrouteservice
fastapi
uvicorn
jinja2
python-multipart
fastapi
uvicorn
jinja2
python-multipart
pydantic
numpy
opencv-python
pillow
aiofiles
requests
insightface
psycopg2-binary